+++ examples/lines_chain_demo/lines_chain_demo.cpp
@@ -90,7 +90,7 @@ public:
     void LoadPng(const draw::Png<Pixel> &png)
     {
         auto maximum = this->demoModel_.color.level.high.GetMaximum();
-        auto scale = static_cast<float>(maximum);
+        auto scale = static_cast<double>(maximum);
 
         // Prevent drawing until new dimensions and source data are
         // synchronized.
